@@ -149,25 +59,123 @@
   "workspaces": [
     "packages/*"
   ],
-  "files": [
-    "lib",
-    "es",
-    "packages",
-    "vetur"
-  ],
-  "main": "lib/index.js",
-  "module": "lib/index.esm.js",
-  "typings": "lib/index.d.ts",
-  "unpkg": "lib/index.js",
-  "style": "lib/theme-default/index.css",
-  "vetur": {
-    "tags": "./lib/tags.json",
-    "attributes": "./lib/attributes.json"
-  },
-  "web-types": "lib/web-types.json",
   "browserslist": [
     "> 1%",
-    "last 2 versions",
-    "not ie <= 10"
-  ]
+    "not ie 11",
+    "not op_mini all"
+  ],
+  "peerDependencies": {
+    "vue": "^3.2.11"
+  },
+  "dependencies": {
+    "@amap/amap-jsapi-loader": "^1.0.1",
+    "heatmap.js": "^2.0.5",
+    "lodash": "^4.17.21",
+    "mitt": "^3.0.0",
+    "proj4": "^2.7.5"
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.15.4",
+    "@babel/core": "^7.15.5",
+    "@babel/plugin-proposal-class-properties": "^7.14.5",
+    "@babel/plugin-transform-runtime": "^7.15.0",
+    "@babel/preset-env": "^7.15.4",
+    "@babel/preset-typescript": "^7.15.0",
+    "@commitlint/cli": "^9.1.2",
+    "@commitlint/config-conventional": "^9.1.2",
+    "@rollup/plugin-alias": "^3.1.5",
+    "@rollup/plugin-commonjs": "^17.1.0",
+    "@rollup/plugin-node-resolve": "^11.2.0",
+    "@rollup/plugin-typescript": "^8.2.0",
+    "@types/gulp": "^4.0.9",
+    "@types/heatmap.js": "^2.0.36",
+    "@types/jest": "^26.0.20",
+    "@types/klaw-sync": "^6.0.1",
+    "@types/lodash": "^4.14.168",
+    "@types/merge-descriptors": "^1.0.1",
+    "@types/proj4": "^2.5.0",
+    "@types/webpack-dev-server": "^4.1.0",
+    "@typescript-eslint/eslint-plugin": "^4.30.0",
+    "@typescript-eslint/parser": "^4.14.2",
+    "@vue/babel-plugin-jsx": "1.0.7",
+    "@vue/compiler-sfc": "3.2.8",
+    "@vue/component-compiler-utils": "^3.2.2",
+    "algoliasearch": "^4.10.5",
+    "babel-jest": "^26.6.3",
+    "babel-loader": "^8.2.2",
+    "babel-plugin-lodash": "^3.3.4",
+    "babel-plugin-module-resolver": "^4.1.0",
+    "babel-preset-vue": "^2.0.2",
+    "chalk": "^4.1.2",
+    "clipboard-copy": "^4.0.1",
+    "components-helper": "^1.0.2",
+    "copy-webpack-plugin": "6.4.1",
+    "coveralls": "^3.1.0",
+    "cp-cli": "^2.0.0",
+    "cross-env": "^7.0.3",
+    "css-loader": "^4.2.2",
+    "css-minimizer-webpack-plugin": "1.1.5",
+    "cz-conventional-changelog": "^3.3.0",
+    "element-plus": "^1.1.0-beta.9",
+    "element-theme-chalk": "^2.15.6",
+    "esbuild": "^0.12.25",
+    "eslint": "^7.19.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-vue": "^7.9.0",
+    "esno": "^0.9.1",
+    "fast-glob": "^3.2.7",
+    "file-loader": "^6.2.0",
+    "file-save": "^0.2.0",
+    "gulp": "^4.0.2",
+    "gulp-typescript": "^6.0.0-alpha.1",
+    "highlight.js": "^11.2.0",
+    "html-webpack-plugin": "4.3.0",
+    "husky": "^4.2.5",
+    "import-from": "^3.0.0",
+    "jest": "^26.6.3",
+    "jest-electron": "^0.1.11",
+    "klaw-sync": "^6.0.0",
+    "lerna": "^3.22.1",
+    "lint-staged": "^10.2.13",
+    "markdown-it": "^11.0.0",
+    "markdown-it-anchor": "^5.3.0",
+    "markdown-it-chain": "^1.3.0",
+    "markdown-it-container": "^3.0.0",
+    "merge-descriptors": "^1.0.1",
+    "mini-css-extract-plugin": "0.11.2",
+    "ora": "^5.4.1",
+    "prettier": "^2.3.2",
+    "progress-bar-webpack-plugin": "^2.1.0",
+    "rimraf": "^3.0.2",
+    "resize-observer-polyfill": "^1.5.1",
+    "rollup": "^2.56.3",
+    "rollup-plugin-css-only": "^3.1.0",
+    "rollup-plugin-esbuild": "^4.5.0",
+    "rollup-plugin-filesize": "^9.1.1",
+    "rollup-plugin-replace": "^2.2.0",
+    "rollup-plugin-terser": "^7.0.2",
+    "rollup-plugin-typescript2": "^0.30.0",
+    "rollup-plugin-vue": "^6.0.0",
+    "sass": "^1.39.2",
+    "sass-loader": "^10.1.1",
+    "style-loader": "^1.3.0",
+    "throttle-debounce": "2.3.0",
+    "through2": "^4.0.2",
+    "transliteration": "^2.2.0",
+    "ts-loader": "^8.0.15",
+    "ts-morph": "^11.0.3",
+    "ts-node": "^10.1.0",
+    "typescript": "^4.4.3",
+    "url-loader": "^4.1.1",
+    "vue": "^3.2.11",
+    "vue-jest": "5.0.0-alpha.5",
+    "vue-loader": "^16.1.2",
+    "vue-router": "^4.0.2",
+    "vue-template-compiler": "^2.6.12",
+    "webpack": "4.44.1",
+    "webpack-bundle-analyzer": "3.9.0",
+    "webpack-cli": "3.3.12",
+    "webpack-dev-server": "3.11.0"
+  }
 }
