@@ -1,12 +1,36 @@
 {
   "name": "@vue-cesium/directives",
   "version": "3.0.1-beta.5.1",
-  "main": "dist/index.js",
+  "description": "VueCesium directives",
+  "main": "index.js",
+  "module": "index.js",
+  "unpkg": "index.js",
+  "jsdelivr": "index.js",
+  "types": "index.d.ts",
   "license": "MIT",
+  "scripts": {
+    "clean": "rm -rf lib",
+    "build": "gulp build"
+  },
+  "bugs": {
+    "url": "https://github.com/zouyaoji/vue-cesium/issues"
+  },
+  "homepage": "https://zouyaoji.top/vue-cesium",
+  "repository": {
+    "type": "git",
+    "url": "git://github.com:zouyaoji/vue-cesium.git"
+  },
   "peerDependencies": {
-    "vue": "^3.0.11"
+    "vue": "^3.2.11"
   },
   "devDependencies": {
-    "@vue/test-utils": "^2.0.0-rc.4"
+    "@types/gulp": "^4.0.9",
+    "@vue/test-utils": "^2.0.0-rc.14",
+    "chalk": "^4.1.2",
+    "gulp": "^4.0.2",
+    "gulp-typescript": "^6.0.0-alpha.1",
+    "sass": "^1.39.2",
+    "ts-node": "^10.1.0",
+    "typescript": "^4.4.3"
   }
 }
