@@ -1,12 +1,29 @@
 {
   "name": "@vue-cesium/utils",
   "version": "3.0.1-beta.13",
-  "main": "dist/index.js",
+  "description": "VueCesium utils.",
+  "main": "index.js",
+  "module": "index.js",
+  "unpkg": "index.js",
+  "jsdelivr": "index.js",
+  "types": "index.d.ts",
   "license": "MIT",
+  "scripts": {
+    "clean": "rm -rf lib",
+    "build": "gulp build"
+  },
+  "bugs": {
+    "url": "https://github.com/zouyaoji/vue-cesium/issues"
+  },
+  "homepage": "https://zouyaoji.top/vue-cesium",
+  "repository": {
+    "type": "git",
+    "url": "git://github.com:zouyaoji/vue-cesium.git"
+  },
   "peerDependencies": {
-    "vue": "^3.0.11"
+    "vue": "^3.2.11"
   },
   "devDependencies": {
-    "@vue/test-utils": "^2.0.0-rc.4"
+    "@vue/test-utils": "^2.0.0-rc.14"
   }
 }
