@@ -1,9 +1,8 @@
+// https://github.com/michael-ciniawsky/postcss-load-config
+
 module.exports = {
-  plugins: {
-    'postcss-import': {},
-    'autoprefixer': {},
-    'postcss-url': {
-      url: 'inline'
-    }
-  },
+  plugins: [
+    // to edit target browsers: use "browserslist" field in package.json
+    require('autoprefixer')
+  ]
 }
