@@ -1,12 +1,14 @@
 export { default as useCommon } from './use-common'
+export { default as useDatasources } from './use-datasources'
 export { default as useEvents } from './use-events'
+export { default as useGeometries } from './use-geometries'
 export { default as useGraphics } from './use-graphics'
 export { default as useHandler } from './use-handler'
 export { default as usePolylineDrawing } from './use-polyline-drawing'
-export { default as useProviders } from './use-providers'
-export { default as useDatasources } from './use-datasources'
-export { default as usePrimitives } from './use-primitives'
-export { default as useGeometries } from './use-geometries'
-export { default as usePrimitiveCollections } from './use-primitive-collections'
 export { default as usePrimitiveCollectionItems } from './use-primitive-collection-items'
+export { default as usePrimitiveCollections } from './use-primitive-collections'
+export { default as usePrimitives } from './use-primitives'
+export { default as useProviders } from './use-providers'
 export { default as useVueCesium } from './use-vue-cesium'
+
+export * from './use-locale'
