@@ -18,7 +18,9 @@ export const i18n = (fn: (...args: any[]) => string) => {
   i18nHandler = fn
 }
 
-function template (str: string, option) {
+export const restoreHandler = () => (i18nHandler = defaultTranslator)
+
+function template(str: string, option) {
   if (!str || !option) return str
 
   return str.replace(/\{(\w+)\}/g, (match, key) => {
@@ -26,9 +28,7 @@ function template (str: string, option) {
   })
 }
 
-export const t = (...args: any[]): string => {
-  if (i18nHandler) return i18nHandler(...args)
-
+const defaultTranslator = (...args: any[]) => {
   const [path, option] = args
   let value
   const array = path.split('.')
@@ -40,7 +40,15 @@ export const t = (...args: any[]): string => {
     if (!value) return ''
     current = value
   }
-  return ''
+}
+
+export const t = (...args: any[]): string => {
+  if (i18nHandler) {
+    const translation = i18nHandler(...args)
+    return translation || defaultTranslator(...args)
+  }
+
+  return defaultTranslator(...args)
 }
 
 export const use = (l: Language): void => {
@@ -50,4 +58,4 @@ export const use = (l: Language): void => {
   }
 }
 
-export default { use, t, i18n }
+export const setLocale = use
