@@ -12,23 +12,33 @@ const highlight = (str, lang) => {
   if (!lang || !hljs.getLanguage(lang)) {
     return '<pre><code class="hljs">' + str + '</code></pre>'
   }
-  const html = hljs.highlight(lang, str, true, undefined).value
+  // const html = hljs.highlight(lang, str, true, undefined).value
+  const html = hljs.highlight(str, {
+    language: lang,
+    ignoreIllegals: true
+  }).value
   return `<pre><code class="hljs language-${lang}">${html}</code></pre>`
 }
 
-config
-  .options.html(true).highlight(highlight).end()
+config.options
+  .html(true)
+  .highlight(highlight)
+  .end()
 
-  .plugin('anchor').use(anchorPlugin, [
+  .plugin('anchor')
+  .use(anchorPlugin, [
     {
       level: 2,
       slugify: slugify,
       permalink: true,
-      permalinkBefore: true,
-    },
-  ]).end()
+      permalinkBefore: true
+    }
+  ])
+  .end()
 
-  .plugin('containers').use(containers).end()
+  .plugin('containers')
+  .use(containers)
+  .end()
 
 const md = config.toMd()
 overWriteFenceRule(md)
